requirements.txt
streamlit
pandas
numpy
matplotlib
openpyxl
sixsigmaspc
seaborn
scipy==1.15.3
statsmodels
ruptures
